Improved feed update behavior
When updating the feeds, we concurrently:
- fetch all users;
- fetch all feeds, which allows us to fetch all stories, add them to
the database and infer from the database response code which article
are new.
When both promises are fulfilled, we update each user concurrently with
the the new unread articles. Thus, Each user object is now only updated
once, which fixes the multiple and concurrent edits done in
74fba1bdcc6cadd4aaa93fcc69cfd996fe25a14c