Improved frontend look with Typicons
Typicons are vector icons  http://typicons.com/

Please run npm install & bower install

1. Typicons is distributed as a css file, but we use scss in our
compile chain. We must then convert the css file into a scss one, with
the help of the new grunt task copy:cssAsScss. See
https://github.com/nex3/sass/issues/556#issuecomment-15718564 and
Gruntfile.js for more details
2. We need to move the Typicons font files to a folder that will be
reachable by the browser, both in a development and production setup.
This will be the "resource" folder, which will be fully copied to the
"dist" folder.
3. As a result, we must change the URL for the Typicons files in the
css source. This is done in the grunt task replace:typicons. See
https://github.com/yoniholmes/grunt-text-replace for documentation
about this module.